REFACTOR: 라벨 출력 부분 추출
